Refact : 추가, 삭제 관련 로직을 컴포넌트화 시킴
