+++ src/type/Comment/comment.ts
@@ -18,7 +18,4 @@ export type ReplyType = {
 
 export type CommentReplySectionDTO = {
   commentId: number;
-  replies: any[];
-  onSubmitReply: (commentId: number, content: string, userId: string) => Promise<void>;
-  onDeleteReply: () => void;
 };
